feature(arena.cc, factoriy_*): feature #9-factory-pattern
Implemented the factory pattern and updated code so it met google style
standards. Backtracked and removed google style fixes (sending pointers
instead of refs,) but I realize now that there is a parse error when the
json file is sent in to initialize the simulation. This is a serious
issue but the unittests work fine.